Add simple syslog test.
This doesn't actaully work right now because musl's syslog
depends on all kinds of things like sockets and being
able to open "/dev/console".

So this is really just a compile and link test.

The indirect dependency on malloc comes from that fact
that it calls gmtime_r which can allocate in JS.